+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Add average slot time (home) ([\#14](https://github.com/forbole/big-dipper-2.0-solana/issues/14))
 - Add latest transactions (home) ([\#17](https://github.com/forbole/big-dipper-2.0-solana/issues/17))
 - Add block list (home) ([\#29](https://github.com/forbole/big-dipper-2.0-solana/issues/29))
+- Add market data on navbar (home) ([\#18](https://github.com/forbole/big-dipper-2.0-solana/issues/18))